Потоки
- Запуск задач заменен на потоки
- Бот вынесен в отдельный файл
