+++ IllsonProject/Cargo.toml
@@ -17,3 +17,4 @@ deadpool-postgres = "0.14.1"
 dotenvy = "0.15"
 envy = "0.4"
 tower-http = { version = "0.6.2", features = ["trace"] }
+crossbeam = "0.8"
